Bouge les parametres dans map-layers pour éviter les imports entre page
