+++ src/pages/map.tsx
@@ -1,40 +1,10 @@
 import { useRouter } from 'next/router';
 
 import IframeWrapper from '@components/IframeWrapper';
-import { type MapLegendFeature } from '@components/Map/components/MapLegendReseaux';
 import Map from '@components/Map/Map';
+import { legendURLKeyToLegendFeature } from '@components/Map/map-layers';
 import useRouterReady from '@hooks/useRouterReady';
 import { createMapConfiguration } from 'src/services/Map/map-configuration';
-
-export const selectableLayers = [
-  {
-    label: 'Les réseaux de chaleur existants',
-    key: 'reseau_chaleur',
-  },
-  {
-    label: 'Les réseaux de chaleur en construction',
-    key: 'futur_reseau',
-  },
-  {
-    label: 'Les périmètres de développement prioritaire',
-    key: 'pdp',
-  },
-  {
-    label: 'Les réseaux de froid',
-    key: 'reseau_froid',
-  },
-] as const;
-
-export type LegendURLKey = (typeof selectableLayers)[number]['key'];
-
-const legendURLKeyToLegendFeature: Record<LegendURLKey | string, MapLegendFeature> = {
-  reseau_chaleur: 'reseauxDeChaleur',
-  futur_reseau: 'reseauxEnConstruction',
-  reseau_froid: 'reseauxDeFroid',
-  pdp: 'zonesDeDeveloppementPrioritaire',
-  raccordementsChaud: 'batimentsRaccordesReseauxChaleur',
-  raccordementsFroid: 'batimentsRaccordesReseauxFroid',
-};
 
 const MapPage = () => {
   const router = useRouter();
